Vertical実装中に伴いパラメータ調整、いらなくなったもの＆警告削除

+++ sln/scene/GamePlayScene.h
@@ -205,7 +205,7 @@ private:
 	float DecelVal = 0.1f;//Œ¸‘¬’l
 
 	//—^‚¦‚éˆĞ—Í
-	int Damage;
+	float Damage;
 
 	std::vector<std::vector<std::string>> csvData;
 	//‰½s–Ú‚Ü‚Åo‚µ‚½‚©
